requirements.txt
streamlit
langchain
python-dotenv
requests
numpy
faiss-cpu
